# docs/v1_swagger.yaml
basePath: /v1
definitions:
  routers.Book:
    properties:
      author:
        type: string
      id:
        type: integer
      title:
        type: string
      year:
        type: integer
    type: object
  routers.NewCollection:
    properties:
      name:
        type: string
      vectorSize:
        type: integer
    type: object
  routers.ToRemoveCollection:
    properties:
      name:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /books:
    get:
      consumes:
      - application/json
      description: get string by ID
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/routers.Book'
            type: array
      summary: Get a list of books in the the store
  /createCollection:
    post:
      consumes:
      - application/json
      description: post string by ID
      parameters:
      - description: New collection description
        in: body
        name: newCollection
        required: true
        schema:
          $ref: '#/definitions/routers.NewCollection'
      produces:
      - application/json
      responses:
        "200":
          description: ok
        "503":
          description: error
      summary: Creates Qdrant Collection
  /removeCollection:
    post:
      consumes:
      - application/json
      description: post string by ID
      parameters:
      - description: Remove collection description
        in: body
        name: removeCollection
        required: true
        schema:
          $ref: '#/definitions/routers.ToRemoveCollection'
      produces:
      - application/json
      responses:
        "200":
          description: ok
        "503":
          description: error
      summary: Creates Qdrant Collection
swagger: "2.0"